vm(genesis-state): tests finished
